feat(radio): handled radio value
