+++ src/components/GroupRadio/GroupRadio.props.ts
@@ -1,5 +1,3 @@
-import { ChangeEventHandler, ReactNode } from 'react'
-
 export type Option = {
   label: string;
   description?: string;
@@ -13,5 +11,5 @@ export type GropupRadioProps = {
   defaultValue: string | number | boolean;
   disabled?: boolean;
   options: Option[];
-  onChange?: ChangeEventHandler;
+  onChange?: (value: string | number | boolean) => void;
 };
